[MOV] website_sale_stock_wishlist: to move
