+++ addons/website_sale_stock/models/__init__.py
@@ -7,6 +7,7 @@ from . import product_feed
 from . import product_product
 from . import product_ribbon
 from . import product_template
+from . import product_wishlist
 from . import res_config_settings
 from . import sale_order
 from . import sale_order_line
