+++ addons/website_sale_stock/models/product_wishlist.py
@@ -1,6 +1,6 @@
-# -*- coding: utf-8 -*-
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
-from odoo import fields, models, api
+
+from odoo import api, fields, models
 
 
 class ProductWishlist(models.Model):
