+++ addons/website_sale_stock/static/src/scss/website_sale_stock_wishlist.scss
