+++ addons/website_sale_stock/templates/wishlist_templates.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <odoo>
-    <template id="product_wishlist" inherit_id="website_sale_wishlist.product_wishlist">
+    <template id="product_wishlist" inherit_id="website_sale.product_wishlist">
         <xpath expr="//t[@t-call='website_sale.products_item']" position="inside">
             <t t-set="product_extra_information">
                 <span
