--- addons/website_sale_stock_wishlist/__manifest__.py
@@ -1,27 +1,0 @@
-# Part of Odoo. See LICENSE file for full copyright and licensing details.
-
-{
-    'name': 'Product Availability Notifications',
-    'category': 'Website/Website',
-    'summary': 'Notify the user when a product is back in stock',
-    'description': """
-Allow the user to select if he wants to receive email notifications when a product of his wishlist gets back in stock.
-    """,
-    'depends': [
-        'website_sale_stock',
-        'website_sale_wishlist',
-    ],
-    'data': [
-        'views/website_sale_stock_wishlist_templates.xml',
-    ],
-    'assets': {
-        'web.assets_frontend': [
-            'website_sale_stock_wishlist/static/src/interactions/**/*',
-            'website_sale_stock_wishlist/static/src/scss/**/*',
-            'website_sale_stock_wishlist/static/src/xml/**/*',
-        ],
-    },
-    'auto_install': True,
-    'author': 'Odoo S.A.',
-    'license': 'LGPL-3',
-}
