--- addons/website_sale_stock_wishlist/i18n/website_sale_stock_wishlist.pot
@@ -1,62 +1,0 @@
-# Translation of Odoo Server.
-# This file contains the translation of the following modules:
-# 	* website_sale_stock_wishlist
-#
-msgid ""
-msgstr ""
-"Project-Id-Version: Odoo Server 19.1a1\n"
-"Report-Msgid-Bugs-To: \n"
-"POT-Creation-Date: 2025-12-16 18:37+0000\n"
-"PO-Revision-Date: 2025-12-16 18:37+0000\n"
-"Last-Translator: \n"
-"Language-Team: \n"
-"MIME-Version: 1.0\n"
-"Content-Type: text/plain; charset=UTF-8\n"
-"Content-Transfer-Encoding: \n"
-"Plural-Forms: \n"
-
-#. module: website_sale_stock_wishlist
-#: model_terms:ir.ui.view,arch_db:website_sale_stock_wishlist.product_wishlist
-msgid ""
-"<i class=\"fa fa-circle text-danger me-1\" aria-hidden=\"true\"/>\n"
-"                    Out of stock"
-msgstr ""
-
-#. module: website_sale_stock_wishlist
-#. odoo-javascript
-#: code:addons/website_sale_stock_wishlist/static/src/xml/product_availability.xml:0
-msgid "Add to wishlist"
-msgstr ""
-
-#. module: website_sale_stock_wishlist
-#. odoo-javascript
-#: code:addons/website_sale_stock_wishlist/static/src/xml/product_availability.xml:0
-msgid "Added to your wishlist"
-msgstr ""
-
-#. module: website_sale_stock_wishlist
-#: model:ir.model.fields,field_description:website_sale_stock_wishlist.field_product_template__display_name
-#: model:ir.model.fields,field_description:website_sale_stock_wishlist.field_product_wishlist__display_name
-msgid "Display Name"
-msgstr ""
-
-#. module: website_sale_stock_wishlist
-#: model:ir.model.fields,field_description:website_sale_stock_wishlist.field_product_template__id
-#: model:ir.model.fields,field_description:website_sale_stock_wishlist.field_product_wishlist__id
-msgid "ID"
-msgstr ""
-
-#. module: website_sale_stock_wishlist
-#: model:ir.model,name:website_sale_stock_wishlist.model_product_template
-msgid "Product"
-msgstr ""
-
-#. module: website_sale_stock_wishlist
-#: model:ir.model,name:website_sale_stock_wishlist.model_product_wishlist
-msgid "Product Wishlist"
-msgstr ""
-
-#. module: website_sale_stock_wishlist
-#: model:ir.model.fields,field_description:website_sale_stock_wishlist.field_product_wishlist__stock_notification
-msgid "Stock Notification"
-msgstr ""
